+++ tests/classic_src/Fields/Interpolation/PPSolveFactoryTest.cpp
@@ -58,10 +58,10 @@ void test_points(int dim, int lower, int upper, std::vector< std::vector<int> > 
       if (upper < 0)
           upper_size = 0;
       // size should be difference in area of the squares
-      EXPECT_EQ(pts.size(), upper_size - lower_size);
+      EXPECT_EQ(pts.size(), (unsigned int)(upper_size - lower_size));
       for (size_t i = 0; i < pts.size(); ++i) {
           // each pts element should have length dim
-          EXPECT_EQ(pts[i].size(), dim);
+          EXPECT_EQ(pts[i].size(), (size_t)dim);
           // each pts element should have indices with lower < size <= upper
           bool in_bounds = true;
           for (int j = 0; j < dim; ++j) {
@@ -154,7 +154,7 @@ TEST_F(PPSolveFactoryTestFixture, TestSolvePolynomialQuadratic) {
     // first check that the polynomial at 0, 0, 0 is the same as ref
     std::vector<double> zero(3, 0.);
     SquarePolynomialVector* test = patch2->getPolynomialVector(&zero[0]);
-    EXPECT_EQ(patch2->getValueDimension(), 2);
+    EXPECT_EQ(patch2->getValueDimension(), (unsigned int)2);
     MMatrix<double> testCoeffs = test->GetCoefficientsAsMatrix();
     std::cout << "Ref" << std::endl;
     std::cout << refCoeffs << std::endl;
@@ -344,7 +344,7 @@ TEST(PPSolveFactoryTest, TestThreeDSolveSinCos) {
             std::stringstream ss;
             ss << "smooth: " << smooth_order << " poly: " << pp_order;
             plot(100, start, end, ppVec.back(), np, ss.str());
-            EXPECT_EQ(ppVec.back()->getValueDimension(), 3);
+            EXPECT_EQ(ppVec.back()->getValueDimension(), (unsigned int)3);
         }
     }
     std::cout << "Testing" << std::endl;
